modification

+++ SapphireHR/Controllers/EmployeeController.cs
@@ -887,7 +887,11 @@ namespace SapphireHR.Web.Controllers
                 var employee = await _employeeService.GetNoTrackingEmployee(id);
                 if (employee != null)
                 {
-                    await _userService.RemoveUser(employee.UserId);
+                    if (!string.IsNullOrEmpty(employee.UserId))
+                    {
+                        await _userService.RemoveUser(employee.UserId);
+                    }
+                    
                     await _employeeService.RemoveEmployee(id);
                 }
                 return Ok();
